// lib/types.ts
export type CardFormData = {
  fullName: string;
  role: string;
  company: string;
  phone: string;
  email: string;
  website: string;
  whatsapp: string;
  instagram: string;
  linkedin: string;
  address: string;
  primaryColor: string;
  secondaryColor: string;
  textColor: string;
  template: "Minimalista" | "Black Premium" | "Corporate" | "Modern Blue" | "White Elegant";
  logoUrl?: string;
};